server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   starlette
    #   watchfiles
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fastapi==0.115.13
    # via -r requirements.in
filelock==3.18.0
    # via torch
fonttools==4.58.4
    # via matplotlib
fsspec==2025.5.1
    # via torch
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
jinja2==3.1.6
    # via torch
kiwisolver==1.4.8
    # via matplotlib
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.3
    # via ultralytics
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.1
    # via ultralytics
opencv-python-headless==4.11.0.86
    # via -r requirements.in
packaging==25.0
    # via
    #   gunicorn
    #   matplotlib
pandas==2.3.0
    # via ultralytics
pillow==11.2.1
    # via
    #   -r requirements.in
    #   matplotlib
    #   torchvision
    #   ultralytics
psutil==7.0.0
    # via ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.0
    # via -r requirements.in
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   ultralytics
    #   uvicorn
redis==6.2.0
    # via -r requirements.in
requests==2.32.4
    # via ultralytics
scipy==1.16.0
    # via ultralytics
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
sympy==1.14.0
    # via ultralytics
tqdm==4.67.1
    # via ultralytics
typing-extensions==4.14.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
ultralytics==8.3.158
    # via -r requirements.in
ultralytics-thop==2.0.14
    # via ultralytics
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.34.3
    # via -r requirements.in
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
